Repeated figures 1 and 2.
Fig 1 now shows the phenotype scale, a correlation matrix and phenotype
differences. The cor matrix is much smaller, and gives correlations within virus
and between viruses.

Fig 2 now gives G and GxE results with an without the top snp as a covariate,
followed by a plot of effect sizes for 10 SNPs.

For the supplementary I need SNP heritability, a table of small loci and linkage
matrix

 On branch dev
 Changes to be committed:
	modified:   02_library/manhattan_plot.R
	modified:   03_scripts/1001genomes_data.R
	modified:   05_figures/fig1/correlation_matrix.R
	modified:   05_figures/fig1/fig1.R
	modified:   05_figures/fig2/fig2.R
	modified:   05_figures/fig2/necrosis-manhattan-plot.R
	new file:   05_figures/fig2/relative_risk/relative_risk.R
	new file:   05_figures/fig2/relative_risk/top_gwas_results.R
	new file:   05_figures/fig2/relative_risk/top_snps.py