Updated supplementary figures
Changed script on SNP heritability to use the kinship matrix with only 1135, and
to remove the joint effects of kinship and major SNP.
Moved the plot for this from figure 2 to the supp figures.

Moved the code to plot LD between SNPs to the supp figures from relative_risk.R.

Fixed paths to GWAS files in supp figures.

 On branch dev
 Changes to be committed:
	modified:   04_analyses/01_snp_heritability/SNP_heritability.Rmd
	modified:   05_figures/fig2/fig2.R
	modified:   05_figures/fig2/relative_risk/relative_risk.R
	modified:   05_figures/supplementary_figures/supplementary_figures.Rmd
	modified:   change_log.md