Symptoms in figure 2
14th April 2023

Rejigged figure 2 to show G and GxE for symptoms and necrosis, plus phenotypes
for the minor alleles, mutants, and TE inserts.

Repeated conditional GWAS for all phenotypes, but moved results to the SI
Merged G and GxE GWAS results for AUDPS and infectivity in the SI.
Reordered stuff in the SI to match the order of the text.

May or may not have messed up removing the relative risk stuff from fig2.

 On branch main
 Changes to be committed:
	new file:   01_data/GWAS_clean_symptoms_mutants.csv
	new file:   01_data/lines_with_TE_insert.txt
	modified:   04_analyses/01_snp_heritability/SNP_heritability.Rmd
	modified:   04_analyses/03_condition_on_top_snp/GWA_condition_on_top_SNP.sh
	modified:   05_figures/fig2/fig2.R
	new file:   05_figures/fig2/mutants.R
	modified:   05_figures/fig2/necrosis-manhattan-plot.R
	new file:   05_figures/fig2/phenotypes_at_major_association.R
	deleted:    05_figures/fig2/relative_risk/output/snp_matrix.csv
	deleted:    05_figures/fig2/relative_risk/output/top_gwas_results.csv
	deleted:    05_figures/fig2/relative_risk/relative_risk.R
	deleted:    05_figures/fig2/relative_risk/top_gwas_results.R
	deleted:    05_figures/fig2/relative_risk/top_snps.py
	new file:   05_figures/fig2/te_insert.R
	modified:   05_figures/supplementary_figures/supplementary_figures.Rmd
	modified:   README.md
	modified:   change_log.md